+++ src/main/java/net/vanillaoutsider/betterdogs/WolfExtensions.java
@@ -50,4 +50,7 @@ public interface WolfExtensions {
 
     boolean betterdogs$isUpForAdoption();
     void betterdogs$setUpForAdoption(boolean adoption);
+
+    long betterdogs$getLastGiftDay();
+    void betterdogs$setLastGiftDay(long day);
 }